Bugfix: Separate GA in/correct click logic from user_answer logic
The condition for determining how to parse out the user_answer to record in the results csv appears to have been repurposed to also determine whether to send a GA correct or incorrect click. This was causing PatMillerErr.answer - that is, 08/16/2018, probably copy pasted from question7 - to be recorded as the user_answer for any correct clicks in the suspicious transactions task.